requirements.txt
sentence-transformers>=2.2.2
faiss-cpu
numpy>=1.24.3
pyyaml>=6.0
pytest>=7.4.0
python-dotenv>=1.0.0
click
# For local LLM support
ollama>=0.1.7
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
# Optional: For API-based models
anthropic>=0.25.0
openai>=1.12.0
google-generativeai
